Use same schema for all view integration tests (#2531)
