+++ tests/chalice/app.py
@@ -1,38 +1,10 @@
-from typing import Any, Optional
-
-import strawberry
 from chalice import Chalice  # type: ignore
 from strawberry.chalice.views import GraphQLView
-from strawberry.types.info import Info
+from tests.views.schema import schema
 
 app = Chalice(app_name="TheStackBadger")
 
 
-@strawberry.type
-class Query:
-    @strawberry.field
-    def greetings(self) -> str:
-        return "hello"
-
-    @strawberry.field
-    def hello(self, name: Optional[str] = None) -> str:
-        return f"Hello {name or 'world'}"
-
-    @strawberry.field
-    def teapot(self, info: Info[Any, None]) -> str:
-        info.context["response"].status_code = 418
-
-        return "🫖"
-
-
-@strawberry.type
-class Mutation:
-    @strawberry.mutation
-    def echo(self, string_to_echo: str) -> str:
-        return string_to_echo
-
-
-schema = strawberry.Schema(query=Query, mutation=Mutation)
 view = GraphQLView(schema=schema, graphiql=True, allow_queries_via_get=True)
 view_no_graphiql = GraphQLView(schema=schema, graphiql=False)
 view_not_get = GraphQLView(schema=schema, graphiql=False, allow_queries_via_get=False)
