+++ tests/channels/test_graphql_transport_ws.py
@@ -18,7 +18,7 @@ from strawberry.subscriptions.protocols.graphql_transport_ws.types import (
     SubscribeMessage,
     SubscribeMessagePayload,
 )
-from tests.channels.schema import schema
+from tests.views.schema import schema
 
 pytestmark = [
     pytest.mark.asyncio,
