+++ tests/django/app/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
 
 from strawberry.django.views import GraphQLView as BaseGraphQLView
-
-from .schema import Query, schema
+from tests.views.schema import Query, schema
 
 
 class GraphQLView(BaseGraphQLView):
