+++ tests/http/clients/asgi.py
@@ -13,9 +13,9 @@ from starlette.websockets import WebSocket
 from strawberry.asgi import GraphQL as BaseGraphQLView
 from strawberry.http import GraphQLHTTPResponse
 from strawberry.types import ExecutionResult
+from tests.views.schema import Query, schema
 
 from ..context import get_context
-from ..schema import Query, schema
 from . import JSON, HttpClient, Response, ResultOverrideFunction
