+++ tests/http/clients/async_flask.py
@@ -7,9 +7,9 @@ from flask import Response as FlaskResponse
 from strawberry.flask.views import AsyncGraphQLView as BaseAsyncGraphQLView
 from strawberry.http import GraphQLHTTPResponse
 from strawberry.types import ExecutionResult
+from tests.views.schema import Query, schema
 
 from ..context import get_context
-from ..schema import Query, schema
 from . import ResultOverrideFunction
 from .flask import FlaskHttpClient
