+++ tests/http/clients/sanic.py
@@ -12,9 +12,9 @@ from strawberry.http import GraphQLHTTPResponse
 from strawberry.http.temporal_response import TemporalResponse
 from strawberry.sanic.views import GraphQLView as BaseGraphQLView
 from strawberry.types import ExecutionResult
+from tests.views.schema import Query, schema
 
 from ..context import get_context
-from ..schema import Query, schema
 from . import JSON, HttpClient, Response, ResultOverrideFunction
