+++ tests/views/schema.py
@@ -1,11 +1,12 @@
 import asyncio
 import typing
 from enum import Enum
-from typing import Optional
 
 from graphql import GraphQLError
 
 import strawberry
+from strawberry.channels.context import StrawberryChannelsContext
+from strawberry.extensions import Extension
 from strawberry.file_uploads import Upload
 from strawberry.permission import BasePermission
 from strawberry.subscriptions.protocols.graphql_transport_ws.types import PingMessage
@@ -15,8 +16,24 @@ from strawberry.types import Info
 class AlwaysFailPermission(BasePermission):
     message = "You are not authorized"
 
-    def has_permission(self, source: typing.Any, info: Info, **kwargs) -> bool:
+    def has_permission(self, source, info, **kwargs) -> bool:
         return False
+
+
+class MyExtension(Extension):
+    def get_results(self):
+        return {"example": "example"}
+
+
+def _read_file(text_file: Upload) -> str:
+    from starlette.datastructures import UploadFile
+
+    # allow to keep this function synchronous, starlette's files have
+    # async methods for reading
+    if isinstance(text_file, UploadFile):
+        text_file = text_file.file._file  # type: ignore
+
+    return text_file.read().decode()
 
 
 @strawberry.enum
@@ -40,53 +57,81 @@ class DebugInfo:
 @strawberry.type
 class Query:
     @strawberry.field
+    def greetings(self) -> str:
+        return "hello"
+
+    @strawberry.field
     def hello(self, name: typing.Optional[str] = None) -> str:
         return f"Hello {name or 'world'}"
 
     @strawberry.field
-    async def async_hello(self, name: str, delay: float = 0) -> str:
+    async def async_hello(
+        self, name: typing.Optional[str] = None, delay: float = 0
+    ) -> str:
         await asyncio.sleep(delay)
         return f"Hello {name or 'world'}"
 
     @strawberry.field(permission_classes=[AlwaysFailPermission])
-    def always_fail(self) -> Optional[str]:
+    def always_fail(self) -> typing.Optional[str]:
         return "Hey"
-
-    @strawberry.field
-    def root_name(root) -> str:
-        return type(root).__name__
 
     @strawberry.field
     async def exception(self, message: str) -> str:
         raise ValueError(message)
         return message
 
+    @strawberry.field
+    def teapot(self, info: Info[typing.Any, None]) -> str:
+        info.context["response"].status_code = 418
+
+        return "🫖"
+
+    @strawberry.field
+    def root_name(self) -> str:
+        return type(self).__name__
+
+    @strawberry.field
+    def value_from_context(self, info: Info) -> str:
+        return info.context["custom_value"]
+
+    @strawberry.field
+    def returns_401(self, info: Info) -> str:
+        response = info.context["response"]
+
+        if hasattr(response, "set_status"):
+            response.set_status(401)
+        else:
+            response.status_code = 401
+
+        return "hey"
+
 
 @strawberry.type
 class Mutation:
     @strawberry.mutation
-    async def hello(self) -> str:
+    def echo(self, string_to_echo: str) -> str:
+        return string_to_echo
+
+    @strawberry.mutation
+    def hello(self) -> str:
         return "strawberry"
 
     @strawberry.mutation
-    async def read_text(self, text_file: Upload) -> str:
-        return (await text_file.read()).decode()
+    def read_text(self, text_file: Upload) -> str:
+        return _read_file(text_file)
 
     @strawberry.mutation
-    async def read_files(self, files: typing.List[Upload]) -> typing.List[str]:
-        contents = []
-        for file in files:
-            content = (await file.read()).decode()
-            contents.append(content)
-        return contents
+    def read_files(self, files: typing.List[Upload]) -> typing.List[str]:
+        return list(map(_read_file, files))
 
     @strawberry.mutation
-    async def read_folder(self, folder: FolderInput) -> typing.List[str]:
-        contents = []
-        for file in folder.files:
-            content = (await file.read()).decode()
-            contents.append(content)
-        return contents
+    def read_folder(self, folder: FolderInput) -> typing.List[str]:
+        return list(map(_read_file, folder.files))
+
+    @strawberry.mutation
+    def match_text(self, text_file: Upload, pattern: str) -> str:
+        text = text_file.read().decode()
+        return pattern if pattern in text else ""
 
 
 @strawberry.type
@@ -150,8 +195,29 @@ class Subscription:
         )
 
     @strawberry.subscription
-    async def connection_params(self, info: Info) -> typing.AsyncGenerator[str, None]:
+    async def listener(
+        self,
+        info: Info[StrawberryChannelsContext, typing.Any],
+        timeout: typing.Optional[float] = None,
+        group: typing.Optional[str] = None,
+    ) -> typing.AsyncGenerator[str, None]:
+        yield info.context.request.channel_name
+
+        async for message in info.context.request.channel_listen(
+            type="test.message",
+            timeout=timeout,
+            groups=[group] if group is not None else [],
+        ):
+            yield message["text"]
+
+    @strawberry.subscription
+    async def connection_params(self, info) -> typing.AsyncGenerator[str, None]:
         yield info.context["connection_params"]["strawberry"]
 
 
-schema = strawberry.Schema(Query, mutation=Mutation, subscription=Subscription)
+schema = strawberry.Schema(
+    query=Query,
+    mutation=Mutation,
+    subscription=Subscription,
+    extensions=[MyExtension],
+)
